+++ src/entities/auto_version.rs
@@ -9,7 +9,7 @@ use crate::i18n;
 /// Rule to extract version from the project automatically.
 #[derive(Deserialize, Serialize, PartialEq, Clone)]
 #[serde(rename_all="snake_case")]
-pub(crate) enum AutoVersionExtractFromRule {
+pub enum AutoVersionExtractFromRule {
   /// Extract from `stdout` of command.
   CmdStdout(CustomCommand),
   /// Extract from plain file.
@@ -18,7 +18,7 @@ pub(crate) enum AutoVersionExtractFromRule {
 
 impl AutoVersionExtractFromRule {
   /// Returns rule type as `&str`.
-  pub(crate) fn type_str(&self) -> &str {
+  pub fn type_str(&self) -> &str {
     match self {
       Self::CmdStdout(_) => i18n::AUTO_VER_CMD_STDOUT,
       Self::PlainFile(_) => i18n::AUTO_VER_PLAIN_FILE,