+++ src/entities/path_type.rs
@@ -3,7 +3,7 @@
 /// Path type.
 /// 
 /// This struct is used only to enable or disable `safe_path` checks inside `add` & `edit` TUIs.
-pub(crate) enum PathType {
+pub enum PathType {
   Absolute,
   Relative,
 }