+++ CHANGELOG.md
@@ -7,6 +7,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.1] - 2025-01-18
+
+### Added
+
+- `deployer docs` command.
+- `tui` feature.
+
+### Changed
+
+- Made all structs and methods public.
+
 ## [1.3.0] - 2025-01-17
 
 ### Added
@@ -306,6 +317,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Build support.
 - TUI.
 
+[1.3.1]:          https://github.com/impulse-sw/deployer/compare/1.3.0...1.3.1
+[1.3.0]:          https://github.com/impulse-sw/deployer/compare/1.2.1...1.3.0
 [1.2.1]:          https://github.com/impulse-sw/deployer/compare/1.2.0...1.2.1
 [1.2.0]:          https://github.com/impulse-sw/deployer/compare/1.1.0-beta-1...1.2.0
 [1.1.0-beta-2]:   https://github.com/impulse-sw/deployer/compare/1.1.0-beta-1...1.1.0-beta-2